♻️ [Refactoring]: toolsディレクトリ内のファイル名を簡潔に変更
- get-プレフィックスを削除し、フォルダ名を考慮した命名に統一
- comment/component/style/version: list.ts
- file: info.ts, nodes.ts
- image: export.ts
- 各index.tsのインポートパスも更新
