+++ src/tools/file/index.ts
@@ -1,8 +1,8 @@
 import type { FigmaApiClient } from '../../api/figma-api-client.js';
 import type { FilesApi } from '../../api/endpoints/files.js';
 import type { FileTools } from './types.js';
-import { createGetFileTool } from './get-file.js';
-import { createGetFileNodesTool } from './get-file-nodes.js';
+import { createGetFileTool } from './info.js';
+import { createGetFileNodesTool } from './nodes.js';
 
 export function createFileTools(apiClient: FigmaApiClient): FileTools {
   const filesApi: FilesApi = apiClient.files;
