+++ src/tools/file/info.ts
