+++ src/tools/image/index.ts
@@ -1,6 +1,6 @@
 import type { FigmaApiClient } from '../../api/figma-api-client.js';
-import { createExportImagesTool } from './export-images.js';
-import type { ImageTool } from './export-images.js';
+import { createExportImagesTool } from './export.js';
+import type { ImageTool } from './export.js';
 
 interface ImageTools {
   exportImages: ImageTool;
